fix: address PR review feedback
- ensureLocks: delete entry on unlock so the per-task mutex map does not grow unbounded
- resolveTaskAgentProfile: return the resolved workflow step alongside the
  profile id so callers (EnsureSession) avoid a redundant lookupWorkflowStep
- httpEnsureTaskSession: route errors through handleNotFound so missing tasks
  return 404 instead of 500
- session_ensure_test: drop the unused v1 import + blank var
- session-launch-service: drop the dead 'none' source variant from the
  EnsureSessionResponse union (backend never returns it)
- find-task: extract shared findTaskInSnapshots utility used by both
  use-task.ts and kanban-with-preview.tsx
